NLP/requirements.txt
gensim==3.8.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
nltk==3.5
numpy==1.18.5
opencv-python==4.3.0.36
pandas==1.0.5
pandocfilters==1.4.2
regex==2020.7.14
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
tensorflow==2.2.0
tokenizers==0.9.4
torch==1.7.1
torchvision==0.8.2
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==4.2.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordfreq==2.3.2
wrapt==1.12.1
wget==3.2
zipp==3.1.0